+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
 
         </activity>
         <activity
-            android:name="com.tenday.feature.exp.lovelguide.LevelGuideActivity"
+            android:name="com.tenday.feature.exp.levelguide.LevelGuideActivity"
             android:exported="true"
             android:theme="@style/Theme.HandsUp">
         </activity>
