+++ feature/exp/src/main/java/com/tenday/feature/exp/levelguide/LevelGuidItem.kt
@@ -1,4 +1,4 @@
-package com.tenday.feature.exp.lovelguide
+package com.tenday.feature.exp.levelguide
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
