+++ feature/exp/src/main/java/com/tenday/feature/exp/levelguide/LevelGuideActivity.kt
@@ -1,4 +1,4 @@
-package com.tenday.feature.exp.lovelguide
+package com.tenday.feature.exp.levelguide
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
