+++ feature/exp/src/main/java/com/tenday/feature/exp/levelguide/LevelGuideScreen.kt
@@ -1,5 +1,5 @@
 
-package com.tenday.feature.exp.lovelguide
+package com.tenday.feature.exp.levelguide
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.ExperimentalFoundationApi
