+++ feature/exp/src/main/java/com/tenday/feature/exp/levelguide/LevelGuideTitleBar.kt
@@ -1,4 +1,4 @@
-package com.tenday.feature.exp.lovelguide
+package com.tenday.feature.exp.levelguide
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
